Remove hint from bundle package
Inside the bundle package there was a Hint printed that was related to
the Copy and Pull command.
